REFACTOR: Extract method that runs the function under test
Adding a helper method that runs the function under test with the instance variables as input, and then assigns the result to another instance variable. The test methods can now call this method without passing any arguments, and interface changes can be done in one place instead of in all test methods.